Make `PrivateKey::from_scalar` constant time
This has no effect on most keys, and simply ensures that the zero scalar
(which is an invalid private key) is not distinguishable.

We also switch to implementing the key validity check by checking that
the public key is not the identity, to more closely match the spec (even
though for ristretto255 this is equivalent to checking for a non-zero
scalar).